Revert grsecurity-3.0-3.2.65-201412310745.patch

+++ fs/sysfs/bin.c
@@ -233,13 +233,13 @@ static int bin_page_mkwrite(struct vm_area_struct *vma, struct vm_fault *vmf)
 	return ret;
 }
 
-static ssize_t bin_access(struct vm_area_struct *vma, unsigned long addr,
-		  void *buf, size_t len, int write)
+static int bin_access(struct vm_area_struct *vma, unsigned long addr,
+		  void *buf, int len, int write)
 {
 	struct file *file = vma->vm_file;
 	struct bin_buffer *bb = file->private_data;
 	struct sysfs_dirent *attr_sd = file->f_path.dentry->d_fsdata;
-	ssize_t ret;
+	int ret;
 
 	if (!bb->vm_ops)
 		return -EINVAL;
